requirements.txt
pandas>=1.5.1
streamlit>=1.15.2
Pillow>=8.0.0
matplotlib>=3.4.0
langchain
openai
streamlit
llama-index==0.8.7
nltk

